requirements.txt
streamlit
pandas
matplotlib.pyplot
seaborn
numpy